moved files into one component folder
